authentication/authorization implementations and debuggage

+++ src/main/java/com/waneib22/taskmanagement/repository/TaskRepository.java
@@ -4,6 +4,8 @@ import com.waneib22.taskmanagement.model.Task;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
+import java.util.Optional;
+
 @Repository
-public interface TaskRepository extends JpaRepository<Task, Long> {
+public interface TaskRepository extends JpaRepository<Task, Integer> {
 }
